+++ CHANGELOG.md
@@ -50,6 +50,8 @@
 - Update to [`libp2p-metrics` `v0.10.0`](misc/metrics/CHANGELOG.md#0100).
 
 - Update to [`libp2p-kad` `v0.41.0`](protocols/kad/CHANGELOG.md#0410).
+- 
+- Update to [`libp2p-identify` `v0.40.0`](protocols/identify/CHANGELOG.md#0400).
 
 # 0.48.0