+++ CHANGELOG.md
@@ -46,6 +46,8 @@
 # 0.49.0 - [unreleased]
 
 - Update to [`libp2p-tcp` `v0.37.0`](transports/tcp/CHANGELOG.md#0370).
+- 
+- Update to [`libp2p-noise` `v0.39.1`](transports/noise/CHANGELOG.md#0391).
 
 - Update to [`libp2p-swarm-derive` `v0.30.1`](swarm-derive/CHANGELOG.md#0301).